refactor: replaced custom timer modal with a form
